Break cache namespace dependency cycles
