+++ ts/packages/cache/src/cache/cache.ts
@@ -38,6 +38,16 @@ import {
 } from "./explainWorkQueue.js";
 import { GrammarStoreImpl } from "./grammarStore.js";
 import { GrammarStore, MatchResult } from "./types.js";
+import {
+    getSchemaNamespaceKeys,
+    splitSchemaNamespaceKey,
+} from "./schemaNamespace.js";
+
+export {
+    getSchemaNamespaceKey,
+    getSchemaNamespaceKeys,
+    splitSchemaNamespaceKey,
+} from "./schemaNamespace.js";
 
 export type ProcessRequestActionResult = {
     explanationResult: ProcessExplanationResult;
@@ -68,39 +78,6 @@ function getFailedResult(message: string): ProcessRequestActionResult {
             },
             elapsedMs: 0,
         },
-    };
-}
-
-export function getSchemaNamespaceKey(
-    name: string,
-    activityName: string | undefined,
-    schemaInfoProvider: SchemaInfoProvider | undefined,
-) {
-    return `${name},${schemaInfoProvider?.getActionSchemaFileHash(name) ?? ""},${activityName ?? ""}`;
-}
-
-// Namespace policy. Combines schema name, file hash, and activity name to indicate enabling/disabling of matching.
-export function getSchemaNamespaceKeys(
-    schemaNames: string[],
-    activityName: string | undefined,
-    schemaInfoProvider: SchemaInfoProvider | undefined,
-) {
-    // Current namespace keys policy is just combining schema name its file hash
-    return schemaNames.map((name) =>
-        getSchemaNamespaceKey(name, activityName, schemaInfoProvider),
-    );
-}
-
-export function splitSchemaNamespaceKey(namespaceKey: string): {
-    schemaName: string;
-    hash: string | undefined;
-    activityName: string | undefined;
-} {
-    const [schemaName, hash, activityName] = namespaceKey.split(",");
-    return {
-        schemaName,
-        hash: hash !== "" ? hash : undefined,
-        activityName: activityName !== "" ? activityName : undefined,
     };
 }
